feat(docker-gc): add support for docker-gc as a container
BREAKING CHANGE:

The project has been renamed to PulseBridge Prompt and the path in which it is saved
now exists at ~/.pulsebridge/prompt. Backward compatibility for restoring bookmarks
is not available. Please restore the bookmarks yourself by copying the bookmarks.sh
file from the ./backup/<DATE>/scripts/bookmarks.sh path to the
$PULSEBRIDGE_PROMPT/scripts/bookmarks.sh path.