feat(bash): install latest bash on Darwin
* install the latest version of bash on Darwin (macOS)
* configure the latest version of bash on Darwin (macOS) as default shell for current user
* remove legacy dotnet SDK from dotnet-install (all platforms)
* build dotnet nuget cache on dotnet-install (all platforms)
* add support for dotnet-cli tab completion
* do not source bashrc from install scripts

BREAKING CHANGE:

This version of prompt will install and configure a new version of bash as the default shell for the user. It does NOT
symlink over /bin/bash. Scripts that do not use /usr/bin/env in the shebang will continue to use the old version of
bash to execute the script, which should maintain backward compatibility on the platform. Please let us know if you have
any issues by creating an aptly named issue on github!