feat: add zsh and dash support to prompt (#32)
* restructure prompt repository to allow switching shells between:
  * sh/dash
  * bash
  * zsh
* move most scripts that are shell agnostic to sh and update shebang
* maintain feature parity between bash / zsh for:
  * prompt
  * color support
  * shell options
  * tab completion (including case-insensitive secondary match)
* update documentation for updated installation
* add `use-bash` command to enable changing the default shell to bash
* add `use-zsh` command to enable changing the default shell to zsh
* add `dark-mode` theme for macOS
* add support for installing older versions of prompt through bootstrap and `update-prompt` commands

NOTES:

Adding support for zsh allows new installations of macOS Catalina to
use the default shell, which was switched to zsh for this release.
Future builds of macOS are likely to drop built-in support for bash due
to the GPLv3 license.

When bootstrapping prompt from bash; prompt will update and use bash.
When bootstrapping prompt from zsh; prompt will update and use zsh.
When bootstrapping from any other shell, prompt will use bash by default.

A command line argument can be used to to select the prompt at install time.
Switching prompts can be done via new `use-bash` and `use-zsh` commands.

BREAKING CHANGE:

While this is not an actual breaking change, quite a bit has changed with
regard to how prompt actually works. We want to notify users that this may
result is some level of bugs as test coverage of scripts like this one is
incredibly difficult.