fix: remove bookmark support
- remove add-bookmark
- remove remove-bookmark
- remove show-bookmarks

NOTE:

The add-bookmark and remove-bookmark scripts have been broken for some
time and no one complained; rather than fix and continue to support a
feature that no one seems to be using, we are removing them.

Folks can add aliases via `edit-bashrc` and `edit-zshrc` if they want
to bookmark locations.