feat: proxy git commands to windows on wsl
NOTE:

This will only work on more recent distributions for WSL but will
dramatically improve performance of the prompt in git paths that are
mounted from the windows file system. The use of the windows version
of git only applies to windows mount path. The local (linux) version of
git is used otherwise.